move private imports from testing into tests/__init__.py
