+++ pandas-stubs/_typing.pyi
@@ -31,6 +31,7 @@ from pandas.core.generic import NDFrame
 from pandas.core.groupby.grouper import Grouper
 from pandas.core.indexes.base import Index
 from pandas.core.series import Series
+from pandas.core.tools.datetimes import FulldatetimeDict
 from typing_extensions import (
     ParamSpec,
     TypeAlias,
@@ -72,22 +73,6 @@ DatetimeDictArg: TypeAlias = (
     Sequence[int] | Sequence[float] | list[str] | tuple[Scalar, ...] | AnyArrayLike
 )
 DictConvertible: TypeAlias = FulldatetimeDict | DataFrame
-
-class YearMonthDayDict(TypedDict, total=True):
-    year: DatetimeDictArg
-    month: DatetimeDictArg
-    day: DatetimeDictArg
-
-class FulldatetimeDict(YearMonthDayDict, total=False):
-    hour: DatetimeDictArg
-    hours: DatetimeDictArg
-    minute: DatetimeDictArg
-    minutes: DatetimeDictArg
-    second: DatetimeDictArg
-    seconds: DatetimeDictArg
-    ms: DatetimeDictArg
-    us: DatetimeDictArg
-    ns: DatetimeDictArg
 
 CorrelationMethod: TypeAlias = (
     Literal["pearson", "kendall", "spearman"]
