+++ tests/test_io.py
@@ -40,7 +40,7 @@ from pandas import (
     read_table,
     read_xml,
 )
-from pandas._testing import ensure_clean
+from pandas.api.typing import JsonReader
 import pytest
 import sqlalchemy
 import sqlalchemy.orm
@@ -51,11 +51,11 @@ from tests import (
     TYPE_CHECKING_INVALID_USAGE,
     WINDOWS,
     check,
+    ensure_clean,
 )
 from tests import NUMPY20  # See https://github.com/PyTables/PyTables/issues/1172
 
 from pandas.io.api import to_pickle
-from pandas.io.json._json import JsonReader
 from pandas.io.parsers import TextFileReader
 from pandas.io.pytables import (
     TableIterator,
